fix: add root conftest so bf_opencore/tests get auth_client and shared fixtures
Move project-level fixtures to root conftest.py, keep minimal tests/conftest.py,
consolidate app fixtures in bf_opencore/tests/conftest.py, update testpaths.
